Recognize POWER11 and future coverage on AIX and Linux
POWER11 will GA next year. It needs to be recognized now in order
to align runtime release with it.

Before OpenJ9 can switch to using OMR's CPU recognition code
completely, it is necessary to recognize furture POWER processors
correctly but as closely as possible for performance purpose
before the relevant macro is officially available from AIX system
header files. A catch-all macro for future processors is defined
as well.

Signed-off-by: Julian <zlwang@ca.ibm.com>